refactor(feed-view): create voting controls component
